requirements.txt
alpaca-py==0.10.0
blinker==1.6.2
boto3==1.28.44
botocore==1.31.44
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
msgpack==1.0.5
numpy==1.25.2
pandas==2.1.0
pydantic==1.10.12
pyngrok==6.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sseclient-py==1.8.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
websockets==10.4
Werkzeug==2.3.7
